+++ src/client.js
@@ -1,14 +1,111 @@
 import 'babel-polyfill'
-import { Provider } from 'react-redux'
-import configureStore from './store/configureStore'
-import { match, browserHistory } from 'react-router'
-import { syncHistoryWithStore } from 'react-router-redux'
-import createRoutes from './routes'
+import 'normalize.css'
 import DeviceProvider from './components/DeviceProvider'
 import MobileDetect from 'mobile-detect'
 import React from 'react'
 import ReactDOM from 'react-dom'
+import ReactGA from 'react-ga'
+import configureStore from './store/configureStore'
+import createRoutes from './routes'
+import { Provider } from 'react-redux'
+import { Router } from 'react-router'
+import { colors, layout, letterSpace, lineHeight, typography } from './themes/common-variables'
+import { injectGlobal } from 'styled-components'
+import { match, browserHistory } from 'react-router'
+import { screen as mq } from './themes/screen'
 import { setupTokenInLocalStorage, deletAuthInfoAction, authUserByTokenAction, keys } from '@twreporter/registration'
+import { syncHistoryWithStore } from 'react-router-redux'
+
+// inject global styles into html
+injectGlobal`
+  html {
+    font-size: ${typography.font.size.base};
+    ::selection {
+      background-color: ${colors.red.lightRed};
+      color: $FFF;
+    }
+  }
+  body {
+    letter-spacing: ${letterSpace.generalLetterSpace};
+
+    line-height: ${lineHeight.lineHeightMedium};
+
+    font-family: "source-han-sans-traditional", "Noto Sans TC", "PingFang TC", "Apple LiGothic Medium", Roboto, "Microsoft JhengHei", "Lucida Grande", "Lucida Sans Unicode", sans-serif;
+
+    abbr[title], abbr[data-original-title] {
+      border-bottom: 0;
+    }
+
+    a {
+      text-decoration: none;
+    }
+
+    .hidden {
+      display: none !important;
+    }
+
+    .container {
+      line-height: ${lineHeight.linHeightLarge};
+    }
+
+    .inner-max {
+      ${mq.desktopAbove`
+        max-width: ${layout.article.innerWidth};
+      `}
+    }
+
+    .outer-max {
+      ${mq.desktopAbove`
+        max-width: ${layout.article.outerWidth};
+      `}
+    }
+
+    .no-hover {
+        border-bottom: 0 !important;
+        &:after {
+            display: none;
+        }
+        &:hover:after {
+            width: 0;
+            display: none;
+        }
+    }
+
+    .text-justify {
+      text-align: justify;
+    }
+
+    .text-center {
+      text-align: center;
+    }
+
+    .center-block {
+      display:block;
+      margin-left:auto;
+      margin-right:auto;
+    }
+
+    .visible-print {
+      display: none;
+    }
+
+    figure, p {
+      margin: 0;
+    }
+
+    @media print {
+      .hidden-print {
+        display: none !important;
+      }
+      .visible-print {
+        display: block !important;
+      }
+      a[href]:after {
+        content: '';
+      }
+    }
+  }
+`
 
 let reduxState
 if (window.__REDUX_STATE__) {
@@ -49,16 +146,29 @@ const device = store.getState().device
 
 const routes = createRoutes(history)
 
-const { pathname, search, hash } = window.location
-const location = `${pathname}${search}${hash}`
+if (typeof window !== 'undefined') {
+  // add Google Analytics
+  ReactGA.initialize('UA-69336956-1')
+  ReactGA.set({ page: window.location.pathname })
+}
+
+
+function scrollAndFireTracking() {
+  if(window) {
+    window.scrollTo(0, 0)
+    // send Google Analytics Pageview event on router changed
+    ReactGA.pageview(window.location.pathname)
+  }
+}
 
 // calling `match` is simply for side effects of
 // loading route/component code for the initial location
-match({ routes, location }, () => {
+// https://github.com/ReactTraining/react-router/blob/v3/docs/guides/ServerRendering.md#async-routes
+match({ history, routes }, (error, redirectLocation, renderProps) => {
   ReactDOM.render((
     <Provider store={store}>
       <DeviceProvider device={device}>
-        { routes }
+        <Router {...renderProps} onUpdate={scrollAndFireTracking}/>
       </DeviceProvider>
     </Provider>
   ), document.getElementById('root'))
@@ -66,13 +176,13 @@ match({ routes, location }, () => {
 
 /* global __DEVTOOLS__ */
 if (__DEVTOOLS__ && !window.__REDUX_DEVTOOLS_EXTENSION__) {
-  match({ routes, location }, () => {
+  match({ history, routes  }, (error, redirectLocation, renderProps) => {
     const DevTools = require('./containers/DevTools').default
     ReactDOM.render((
       <Provider store={store}>
         <DeviceProvider device={device}>
           <div>
-            { routes }
+            <Router {...renderProps}/>
             <DevTools />
           </div>
         </DeviceProvider>