+++ src/constants/customized-article-path.js
@@ -1,0 +1,9 @@
+// Create customized route
+// 1. Declare your own variables
+// 2. Add them to both exported array "paths" and default object
+
+const HIGH_RISK_YOUTH = 'high-risk-youth-left-in-relic'
+
+export default {
+  HIGH_RISK_YOUTH
+}