+++ src/helpers/Html.js
@@ -2,16 +2,20 @@
 import Helmet from 'react-helmet'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
+import map from 'lodash/map'
 import serialize from 'serialize-javascript'
+
+const _ = {
+  map
+}
 
 export default class Html extends PureComponent {
   static propTypes = {
-    assets: PropTypes.object,
+    assets: PropTypes.object.isRequired,
     component: PropTypes.node,
-    store: PropTypes.object,
-    styleTags: PropTypes.string
+    store: PropTypes.object.isRequired,
+    styleTags: PropTypes.string.isRequired
   }
-
   render() {
     const { assets, content, store, styleTags } = this.props
     const head = Helmet.rewind()
@@ -39,8 +43,8 @@ export default class Html extends PureComponent {
           <link href="https://www.twreporter.org/images/icon-hires.png" rel="icon" sizes="192x192" />
           <link href="https://www.twreporter.org/images/icon-normal.png" rel="icon" sizes="128x128" />
           <link href="/asset/favicon.png"  rel="shortcut icon" />
-          {Object.keys(assets.styles).map((style, key) =>
-            <link href={assets.styles[style]} key={key} media="all"
+          {_.map(assets.stylesheets, (stylesheet, key) =>
+            <link href={stylesheet} key={'stylesheet' + key} media="all"
               rel="stylesheet" type="text/css" charSet="UTF-8"/>
           )}
           <style dangerouslySetInnerHTML={{ __html: styleTags }} />
@@ -49,7 +53,7 @@ export default class Html extends PureComponent {
           <div id="root" dangerouslySetInnerHTML={{ __html: content }} />
           <script defer src="https://cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.zh-Hant-TW" />
           <script dangerouslySetInnerHTML={{ __html: `window.__REDUX_STATE__=${serialize(store.getState())};` }} charSet="UTF-8"/>
-          <script src={assets.javascript.main} charSet="UTF-8"/>
+          <script src={assets.javascripts.main} charSet="UTF-8"/>
 
           {/* <!-- Load typekit fonts for twreporter.org domain--> */}
           <script