+++ src/routes/index.js
@@ -3,188 +3,140 @@ if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
 
 import App from '../containers/App'
 import React from 'react'
-import ReactGA from 'react-ga'
 import { Route, Router, browserHistory } from 'react-router'
 
-if (typeof window !== 'undefined') {
-  // add Google Analytics
-  ReactGA.initialize('UA-69336956-1')
-  ReactGA.set({ page: window.location.pathname })
-}
-
-// The variable host is declared for @twreporter/registration
-export const ACTIVATE_PAGE_PATH = 'activate'
-
-function scrollAndFireTracking() {
-  if(window) {
-    window.scrollTo(0, 0)
-    // send Google Analytics Pageview event on router changed
-    ReactGA.pageview(window.location.pathname)
+function loadRoute(cb) {
+  return (module) => {
+    cb(null, module.default || module)
   }
 }
 
-/* Code splitting and dynamic loading bundles by webpack require.ensure.
- * require.ensure is defined only in webpack, so the first line is a polyfill for node environment.
- * Due to webpack limitation, the following code is kind of duplicated.
- * Currently, I have no way to reduce the codes.
- * If you know how to reduce the duplicated codes, PR is welcome.
+// The variable is declared for @twreporter/registration
+export const ACTIVATE_PAGE_PATH = 'activate'
+
+function errorLoading(err) {
+  console.error('Err to load module:', err) //eslint-disable-line
+}
+
+/**
+ * createRoutes
+ * Code splitting and dynamic loading bundles by webpack require.ensure.
+ * require.ensure is defined only in webpack, so the first line in this file is a polyfill for node environment.
+ *
+ * The following `import` in the `getComponent` function will be transpiled to `require.ensure` by babel
+ * through `babel-plugin-system-import-transformer` plugin.
+ *
+ * @param {object} history default is react-router.browserHistory
+ * @returns {object} React Router component
  */
-export default function (history = browserHistory) {
+export default function createRoutes(history = browserHistory) {
   return (
-    <Router history={history} onUpdate={scrollAndFireTracking} >
+    <Router history={history}>
       <Route
         path="/topics/:slug"
-        getComponent={(nextStateWithLocation, cb) => {
-          require.ensure([], (require) => {
-            const module = require('../containers/TopicLandingPage')
-            cb(null, module.default || module)
-          })
+        getComponent={(location, cb) => {
+          import('../containers/TopicLandingPage').then(loadRoute(cb)).catch(errorLoading)
         }}
       />
       <Route
         path="/"
-        getComponent={(nextStateWithLocation, cb) => {
-          require.ensure([], (require) => {
-            const module = require('../containers/Home')
-            cb(null, module.default || module)
-          })
+        getComponent={(location, cb) => {
+          import('../containers/Home').then(loadRoute(cb)).catch(errorLoading)
         }}
       />
       <Route path="/" component={App}>
         <Route
           path="topics"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Topics')
-              cb(null, module.default || module)
-            })
-          }}
-        />
-        <Route
-          path="category/:category"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Category')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Topics').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="categories/:category"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Category')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Category').then(loadRoute(cb)).catch(errorLoading)
+          }}
+        />
+        <Route
+          path="category/:category"
+          getComponent={(location, cb) => {
+            import('../containers/Category').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="topic/:topicId"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Topic')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Topic').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="tag/:tagId"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Tag')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Tag').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="photography"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Photography')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Photography').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="search"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Search')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Search').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="a/:slug"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Article')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Article').then(loadRoute(cb)).catch(errorLoading)
           }} />
         <Route
           path="author/:authorId"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Author')
-              cb(null, module.default || module)
-            })
-          }}
-        />
+          getComponent={(location, cb) => {
+            import('../containers/Author').then(loadRoute(cb)).catch(errorLoading)
+          }} />
         <Route
           path="authors"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/AuthorsList')
-              cb(null, module.default || module)
-            })
-          }}
-        />
+          getComponent={(location, cb) => {
+            import('../containers/AuthorsList').then(loadRoute(cb)).catch(errorLoading)
+          }} />
         <Route
           path={ACTIVATE_PAGE_PATH}
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/Activation')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/Activation').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="signin"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/sign-in')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/sign-in').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="signup"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/sign-up')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/sign-up').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="confirm"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const module = require('../containers/confirm-after-sign')
-              cb(null, module.default || module)
-            })
+          getComponent={(location, cb) => {
+            import('../containers/confirm-after-sign').then(loadRoute(cb)).catch(errorLoading)
           }}
         />
         <Route
           path="bookmarks(/:pageNumber)"
           redirectPath="/signin"
-          getComponent={(nextStateWithLocation, cb) => {
-            require.ensure([], (require) => {
-              const subModule = require('../containers/BookmarkList')
-              const AuthScreen = require('@twreporter/registration').AuthenticationScreen
-              cb(null, AuthScreen(subModule.default))
-            })
+          getComponent={(location, cb) => {
+            Promise.all([
+              import('twreporter/registration'),
+              import('../containers/BookmarkList')
+            ]).then(([ regModule, bookmarkModule ]) => {
+              const AuthScreen = regModule.AuthenticationScreen
+              cb(null, AuthScreen(bookmarkModule.default))
+            }).catch(errorLoading)
           }}
         />
       </Route>